+++ lib/commands/generate.js
@@ -3,7 +3,7 @@
 const path = require('path');
 const { default: chalk } = require('chalk');
 const Command = require('../models/command');
-const Blueprint = require('../models/blueprint');
+const Blueprint = require('@ember-tooling/blueprint-model');
 const mergeBlueprintOptions = require('../utilities/merge-blueprint-options');
 const merge = require('lodash/merge');
 const reject = require('lodash/reject');
