+++ lib/models/command.js
@@ -6,7 +6,7 @@ const path = require('path');
 const isGitRepo = require('is-git-url');
 const camelize = require('ember-cli-string-utils').camelize;
 const getCallerFile = require('get-caller-file');
-const printCommand = require('../utilities/print-command');
+const printCommand = require('@ember-tooling/blueprint-model/utilities/print-command');
 const union = require('lodash/union');
 const defaults = require('lodash/defaults');
 const uniq = require('lodash/uniq');
