+++ lib/models/package-info-cache/package-info.js
@@ -4,7 +4,7 @@ const path = require('path');
 const ErrorList = require('./error-list');
 const Errors = require('./errors');
 const AddonInfo = require('../addon-info');
-const isAddon = require('../../utilities/is-addon');
+const isAddon = require('@ember-tooling/blueprint-model/utilities/is-addon');
 const isEngine = require('../../utilities/is-engine');
 const isLazyEngine = require('../../utilities/is-lazy-engine');
 const logger = require('heimdalljs-logger')('ember-cli:package-info-cache:package-info');
