+++ lib/tasks/generate-from-blueprint.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Blueprint = require('../models/blueprint');
+const Blueprint = require('@ember-tooling/blueprint-model');
 const Task = require('../models/task');
 const parseOptions = require('../utilities/parse-options');
 const logger = require('heimdalljs-logger')('ember-cli:generate-from-blueprint');
