+++ lib/tasks/install-blueprint.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const fs = require('fs-extra');
-const Blueprint = require('../models/blueprint');
+const Blueprint = require('@ember-tooling/blueprint-model');
 const Task = require('../models/task');
 const os = require('os');
 const path = require('path');
