+++ lib/utilities/merge-blueprint-options.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const SilentError = require('silent-error');
-const Blueprint = require('../models/blueprint');
+const Blueprint = require('@ember-tooling/blueprint-model');
 
 /*
  * Helper for commands that use a blueprint to merge the blueprint's options
