+++ packages/blueprint-model/package.json
@@ -8,12 +8,24 @@
   },
   "license": "MIT",
   "dependencies": {
+    "babel-remove-types": "^2.0.0",
     "chalk": "^5.6.2",
+    "content-tag": "^4.0.0",
+    "core-object": "^3.1.5",
     "diff": "^8.0.4",
+    "ember-cli-normalize-entity-name": "^1.0.0",
+    "ember-cli-string-utils": "^1.1.0",
+    "fs-extra": "^11.3.4",
+    "heimdalljs-logger": "^0.1.10",
+    "inflection": "^3.0.2",
     "isbinaryfile": "^5.0.7",
     "lodash": "^4.17.23",
+    "markdown-it": "^14.1.0",
+    "markdown-it-terminal": "^0.4.0",
+    "minimatch": "^10.1.1",
     "promise.hash.helper": "^1.0.8",
     "quick-temp": "^0.1.9",
-    "silent-error": "^1.1.1"
+    "silent-error": "^1.1.1",
+    "walk-sync": "^4.0.0"
   }
 }
