+++ packages/blueprint-model/utilities/clean-remove.js
