+++ packages/blueprint-model/utilities/insert-into-file.js
