+++ packages/blueprint-model/utilities/is-addon.js
