+++ packages/blueprint-model/utilities/markdown-color.js
