+++ packages/blueprint-model/utilities/sequence.js
