+++ packages/blueprint-model/utilities/walk-up-path.js
