+++ tests/acceptance/addon-dummy-generate-test.js
@@ -2,7 +2,7 @@
 
 const ember = require('../helpers/ember');
 let root = process.cwd();
-const Blueprint = require('../../lib/models/blueprint');
+const Blueprint = require('@ember-tooling/blueprint-model');
 const BlueprintNpmTask = require('ember-cli-internal-test-helpers/lib/helpers/disable-npm-on-blueprint');
 const tmp = require('tmp-promise');
