+++ tests/acceptance/addon-generate-test.js
@@ -4,7 +4,7 @@ const ember = require('../helpers/ember');
 const path = require('path');
 const fs = require('fs-extra');
 let root = process.cwd();
-const Blueprint = require('../../lib/models/blueprint');
+const Blueprint = require('@ember-tooling/blueprint-model');
 const BlueprintNpmTask = require('ember-cli-internal-test-helpers/lib/helpers/disable-npm-on-blueprint');
 const tmp = require('tmp-promise');
