+++ tests/acceptance/in-option-generate-test.js
@@ -4,7 +4,7 @@ const ember = require('../helpers/ember');
 const fs = require('fs-extra');
 const path = require('path');
 let root = process.cwd();
-const Blueprint = require('../../lib/models/blueprint');
+const Blueprint = require('@ember-tooling/blueprint-model');
 const BlueprintNpmTask = require('ember-cli-internal-test-helpers/lib/helpers/disable-npm-on-blueprint');
 const tmp = require('tmp-promise');
 const initApp = require('../helpers/init-app');
