+++ tests/acceptance/pods-generate-test.js
@@ -7,7 +7,7 @@ const path = require('path');
 let root = process.cwd();
 const tmp = require('tmp-promise');
 
-const Blueprint = require('../../lib/models/blueprint');
+const Blueprint = require('@ember-tooling/blueprint-model');
 const BlueprintNpmTask = require('ember-cli-internal-test-helpers/lib/helpers/disable-npm-on-blueprint');
 
 const { expect } = require('chai');
