+++ tests/acceptance/typescript-destroy-test.js
@@ -5,7 +5,7 @@ const fs = require('fs-extra');
 let root = process.cwd();
 const tmp = require('tmp-promise');
 
-const Blueprint = require('../../lib/models/blueprint');
+const Blueprint = require('@ember-tooling/blueprint-model');
 const BlueprintNpmTask = require('ember-cli-internal-test-helpers/lib/helpers/disable-npm-on-blueprint');
 
 const { expect } = require('chai');
