+++ tests/integration/utilities/clean-remove-test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { expect } = require('chai');
-const cleanRemove = require('../../../lib/utilities/clean-remove');
+const cleanRemove = require('@ember-tooling/blueprint-model/utilities/clean-remove');
 const os = require('os');
 const path = require('path');
 const fs = require('fs-extra');
