+++ tests/unit/commands/generate-test.js
@@ -6,7 +6,7 @@ const commandOptions = require('../../factories/command-options');
 const processHelpString = require('../../helpers/process-help-string');
 const MockProject = require('../../helpers/mock-project');
 const Task = require('../../../lib/models/task');
-const Blueprint = require('../../../lib/models/blueprint');
+const Blueprint = require('@ember-tooling/blueprint-model');
 const GenerateCommand = require('../../../lib/commands/generate');
 const td = require('testdouble');
 const ROOT = process.cwd();
