+++ tests/unit/commands/init-test.js
@@ -5,7 +5,7 @@ const os = require('os');
 const path = require('path');
 const { expect } = require('chai');
 const MockUI = require('console-ui/mock');
-const Blueprint = require('../../../lib/models/blueprint');
+const Blueprint = require('@ember-tooling/blueprint-model');
 const Project = require('../../../lib/models/project');
 const Task = require('../../../lib/models/task');
 const InitCommand = require('../../../lib/commands/init');
